# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "slate-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/external-holders": {
      "get": {
        "tags": [
          "external-holders"
        ],
        "summary": "List External Holders",
        "operationId": "list_external_holders_external_holders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExternalHolderOut"
                  },
                  "type": "array",
                  "title": "Response List External Holders External Holders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "external-holders"
        ],
        "summary": "Create External Holder",
        "operationId": "create_external_holder_external_holders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalHolderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalHolderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/external-holders/balance": {
      "get": {
        "tags": [
          "external-holders"
        ],
        "summary": "List External Holder Balances",
        "description": "Quanto a household deve a cada titular externo. Derivado, nunca armazenado:\nmarcar a entry como `pago` é o registro do reembolso, sem transferência à parte.\nTitular sem nada em aberto vem com saldo zero.",
        "operationId": "list_external_holder_balances_external_holders_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExternalHolderBalanceOut"
                  },
                  "type": "array",
                  "title": "Response List External Holder Balances External Holders Balance Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/external-holders/{holder_id}": {
      "get": {
        "tags": [
          "external-holders"
        ],
        "summary": "Get External Holder",
        "operationId": "get_external_holder_external_holders__holder_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "holder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Holder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalHolderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "external-holders"
        ],
        "summary": "Update External Holder",
        "operationId": "update_external_holder_external_holders__holder_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "holder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Holder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalHolderIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalHolderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "external-holders"
        ],
        "summary": "Delete External Holder",
        "operationId": "delete_external_holder_external_holders__holder_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "holder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Holder Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/commitments": {
      "post": {
        "tags": [
          "commitments"
        ],
        "summary": "Create Commitment",
        "description": "Cria o commitment e materializa as entries na mesma transação.\n\nParcelamento e avulso saem com todas as entries; recorrente sai com as do horizonte\nrolante (`hoje + 24 meses`, no fuso do job de horizonte), limitadas pelo `end_date`.",
        "operationId": "create_commitment_commitments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentWithEntries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/commitments/active": {
      "get": {
        "tags": [
          "commitments"
        ],
        "summary": "List Active Commitments",
        "operationId": "list_active_commitments_commitments_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActiveCommitmentOut"
                  },
                  "type": "array",
                  "title": "Response List Active Commitments Commitments Active Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/commitments/{commitment_id}": {
      "delete": {
        "tags": [
          "commitments"
        ],
        "summary": "Delete Commitment",
        "description": "Sem `keep_paid`, apaga o commitment e todas as entries, inclusive as pagas.\n\n`keep_paid=true` é o cancelamento: preserva as entries pagas e confirmadas, remove as\nprevistas e o commitment passa a `cancelled`. Só commitment `active` é cancelável.",
        "operationId": "delete_commitment_commitments__commitment_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          },
          {
            "name": "keep_paid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Keep Paid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "keep_paid=true num commitment que não está active"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActiveCommitmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "installment",
              "recurring",
              "single"
            ],
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "purchase_date": {
            "type": "string",
            "format": "date",
            "title": "Purchase Date"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "installment_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Count"
          },
          "recurring_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurring Amount"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "settled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "outstanding_balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Outstanding Balance"
          },
          "payoff_month": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payoff Month"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "category_id",
          "kind",
          "description",
          "purchase_date",
          "total_amount",
          "installment_count",
          "recurring_amount",
          "end_date",
          "status",
          "created_at",
          "outstanding_balance",
          "payoff_month"
        ],
        "title": "ActiveCommitmentOut"
      },
...
      "CommitmentCreate": {
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "installment",
              "recurring",
              "single"
            ],
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Description"
          },
          "purchase_date": {
            "type": "string",
            "format": "date",
            "title": "Purchase Date"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "installment_count": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 360.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Count"
          },
          "recurring_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurring Amount"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "category_id",
          "kind",
          "description",
          "purchase_date"
        ],
        "title": "CommitmentCreate"
      },
      "CommitmentWithEntries": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "installment",
              "recurring",
              "single"
            ],
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "purchase_date": {
            "type": "string",
            "format": "date",
            "title": "Purchase Date"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          },
          "installment_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Count"
          },
          "recurring_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurring Amount"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "settled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "category_id",
          "kind",
          "description",
          "purchase_date",
          "total_amount",
          "installment_count",
          "recurring_amount",
          "end_date",
          "status",
          "created_at",
          "entries"
        ],
        "title": "CommitmentWithEntries"
      },
...
      "ExternalHolderBalanceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance",
            "description": "Soma das entries não pagas das contas do titular; zero se não há nada aberto"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "balance"
        ],
        "title": "ExternalHolderBalanceOut"
      },
      "ExternalHolderIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ExternalHolderIn"
      },
      "ExternalHolderOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "ExternalHolderOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}